@@ -1,6 +1,6 @@
 {
   "name": "if-tsb",
-  "version": "0.2.14",
+  "version": "0.2.15",
   "description": "Insanely Fast TypeScript Bundler",
   "main": "./index.bundle.js",
   "bin": {
@@ -26,13 +26,15 @@
   "dependencies": {
     "@types/glob-to-regexp": "^0.4.0",
     "colors": "^1.4.0",
+    "glob": "^7.1.7",
     "glob-to-regexp": "^0.4.1",
     "source-map": "~0.6.1",
-    "tslib": "^2.1.0",
-    "typescript": "^4.1.3"
+    "tslib": "^2.2.0",
+    "typescript": "^4.2.4"
   },
   "devDependencies": {
     "@types/find-cache-dir": "^3.2.0",
+    "@types/glob": "^7.1.3",
     "@types/node": "^6.14.13",
     "ts-node": "^9.1.1"
   }
